adds a unit test for syntactic usages
