+++ internal/codeintel/codenav/BUILD.bazel
@@ -63,6 +63,7 @@ go_test(
     timeout = "short",
     srcs = [
         "gittree_translator_test.go",
+        "helpers_test.go",
         "mapped_index_test.go",
         "scip_utils_test.go",
         "service_closest_uploads_test.go",
@@ -99,6 +100,7 @@ go_test(
         "//internal/search/streaming",
         "//internal/types",
         "//lib/codeintel/precise",
+        "//lib/errors",
         "@com_github_google_go_cmp//cmp",
         "@com_github_life4_genesis//slices",
         "@com_github_sourcegraph_go_diff//diff",
